Refactor settings navigation and OSS license integration
Moved settings navigation logic from feature/ui-setting to individual app hosts (Cnubus, MyGrade, Nanda) and removed SettingsNavigation.kt. Integrated Google Play OSS licenses into app modules. Updated TimberInitializer to use FirebaseInitializer from core.data. Removed redundant FirebaseInitializer and debug AndroidManifest files.
