+++ app-my-grade/src/main/java/com/keelim/mygrade/ui/MyGradeHost.kt
@@ -2,6 +2,8 @@
 
 package com.keelim.mygrade.ui
 
+import android.content.Context
+import android.content.Intent
 import androidx.compose.animation.ContentTransform
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
@@ -9,13 +11,17 @@ import androidx.compose.animation.scaleOut
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.core.net.toUri
 import androidx.lifecycle.viewmodel.navigation3.rememberViewModelStoreNavEntryDecorator
+import androidx.navigation3.runtime.EntryProviderBuilder
 import androidx.navigation3.runtime.entry
 import androidx.navigation3.runtime.entryProvider
 import androidx.navigation3.runtime.rememberSavedStateNavEntryDecorator
 import androidx.navigation3.ui.NavDisplay
+import com.google.android.gms.oss.licenses.OssLicensesMenuActivity
 import com.keelim.composeutil.AppState
 import com.keelim.composeutil.rememberMutableStateListOf
 import com.keelim.core.navigation.AppRoute
@@ -34,8 +40,14 @@ import com.keelim.mygrade.ui.screen.task.chart.TaskChartRoute
 import com.keelim.mygrade.ui.screen.timer.history.TimerHistoryRoute
 import com.keelim.mygrade.ui.screen.word.show.WordShowRoute
 import com.keelim.mygrade.ui.screen.word.write.WordWriteRoute
+import com.keelim.setting.screen.admin.AdminRoute
+import com.keelim.setting.screen.alarm.AlarmRoute
 import com.keelim.setting.screen.event.EventRoute
-import com.keelim.setting.screen.settings.settingsEntry
+import com.keelim.setting.screen.faq.FaqRoute
+import com.keelim.setting.screen.lab.LabRoute
+import com.keelim.setting.screen.notification.NotificationRoute
+import com.keelim.setting.screen.settings.SettingsRoute
+import com.keelim.setting.screen.theme.ThemeRoute
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
@@ -177,3 +189,60 @@ fun MyGradeHost(
         },
     )
 }
+
+@Composable
+fun EntryProviderBuilder<Any>.settingsEntry(
+    backStack: SnapshotStateList<Any>,
+    context: Context
+) {
+    entry<FeatureRoute.Settings> {
+        SettingsRoute(
+            onThemeChangeClick = { backStack.add(FeatureRoute.Theme) },
+            onNotificationsClick = {
+                backStack.add(FeatureRoute.Notification)
+            },
+            onAlarmsClick = {
+                backStack.add(FeatureRoute.Alarm)
+            },
+            onFaqClick = {
+                backStack.add(FeatureRoute.Faq)
+            },
+            onOpenSourceClick = {
+                context.startActivity(Intent(context, OssLicensesMenuActivity::class.java))
+            },
+            onLabClick = {
+                backStack.add(FeatureRoute.Lab)
+            },
+            onAppUpdateClick = {
+                context.startActivity(
+                    Intent(
+                        Intent.ACTION_VIEW,
+                        "https://play.google.com/store/apps/details?id=${context.packageName}".toUri(),
+                    ),
+                )
+            },
+            onAdminClick = {
+                backStack.add(FeatureRoute.Admin)
+            },
+        )
+    }
+    entry<FeatureRoute.Faq> {
+        FaqRoute()
+    }
+    entry<FeatureRoute.Theme> {
+        ThemeRoute()
+    }
+    entry<FeatureRoute.Notification> {
+        NotificationRoute()
+    }
+    entry<FeatureRoute.Lab> {
+        LabRoute()
+    }
+    entry<FeatureRoute.Alarm> {
+        AlarmRoute()
+    }
+    entry<FeatureRoute.Admin> {
+        AdminRoute()
+    }
+}
+
