+++ app-nanda/src/main/java/com/keelim/nandadiagnosis/ui/NandaHost.kt
@@ -2,6 +2,7 @@
 
 package com.keelim.nandadiagnosis.ui
 
+import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import androidx.compose.animation.ContentTransform
@@ -12,14 +13,17 @@ import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.SheetState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.core.net.toUri
 import androidx.lifecycle.viewmodel.navigation3.rememberViewModelStoreNavEntryDecorator
+import androidx.navigation3.runtime.EntryProviderBuilder
 import androidx.navigation3.runtime.entry
 import androidx.navigation3.runtime.entryProvider
 import androidx.navigation3.runtime.rememberSavedStateNavEntryDecorator
 import androidx.navigation3.ui.NavDisplay
+import com.google.android.gms.oss.licenses.OssLicensesMenuActivity
 import com.keelim.composeutil.AppState
 import com.keelim.composeutil.rememberMutableStateListOf
 import com.keelim.core.navigation.AppRoute
@@ -34,8 +38,14 @@ import com.keelim.nandadiagnosis.ui.screen.food.overview.FoodRoute
 import com.keelim.nandadiagnosis.ui.screen.main.MainBottomSheet
 import com.keelim.nandadiagnosis.ui.screen.nutrient.NutrientRoute
 import com.keelim.nandadiagnosis.ui.screen.nutrient.timer.NutrientTimerRoute
+import com.keelim.setting.screen.admin.AdminRoute
+import com.keelim.setting.screen.alarm.AlarmRoute
 import com.keelim.setting.screen.event.EventRoute
-import com.keelim.setting.screen.settings.settingsEntry
+import com.keelim.setting.screen.faq.FaqRoute
+import com.keelim.setting.screen.lab.LabRoute
+import com.keelim.setting.screen.notification.NotificationRoute
+import com.keelim.setting.screen.settings.SettingsRoute
+import com.keelim.setting.screen.theme.ThemeRoute
 import com.keelim.web.navigateToWebModule
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
@@ -160,3 +170,60 @@ fun NandaHost(
         },
     )
 }
+
+@Composable
+private fun EntryProviderBuilder<Any>.settingsEntry(
+    backStack: SnapshotStateList<Any>,
+    context: Context
+) {
+    entry<FeatureRoute.Settings> {
+        SettingsRoute(
+            onThemeChangeClick = { backStack.add(FeatureRoute.Theme) },
+            onNotificationsClick = {
+                backStack.add(FeatureRoute.Notification)
+            },
+            onAlarmsClick = {
+                backStack.add(FeatureRoute.Alarm)
+            },
+            onFaqClick = {
+                backStack.add(FeatureRoute.Faq)
+            },
+            onOpenSourceClick = {
+                context.startActivity(Intent(context, OssLicensesMenuActivity::class.java))
+            },
+            onLabClick = {
+                backStack.add(FeatureRoute.Lab)
+            },
+            onAppUpdateClick = {
+                context.startActivity(
+                    Intent(
+                        Intent.ACTION_VIEW,
+                        "https://play.google.com/store/apps/details?id=${context.packageName}".toUri(),
+                    ),
+                )
+            },
+            onAdminClick = {
+                backStack.add(FeatureRoute.Admin)
+            },
+        )
+    }
+    entry<FeatureRoute.Faq> {
+        FaqRoute()
+    }
+    entry<FeatureRoute.Theme> {
+        ThemeRoute()
+    }
+    entry<FeatureRoute.Notification> {
+        NotificationRoute()
+    }
+    entry<FeatureRoute.Lab> {
+        LabRoute()
+    }
+    entry<FeatureRoute.Alarm> {
+        AlarmRoute()
+    }
+    entry<FeatureRoute.Admin> {
+        AdminRoute()
+    }
+}
+
