+++ core/common-android/src/release/java/com/keelim/commonAndroid/initialize/TimberInitializer.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import androidx.startup.Initializer
 import com.keelim.commonAndroid.util.CrashlyticsTree
 import timber.log.Timber
+import com.keelim.core.data.initialize.FirebaseInitializer
 
 class TimberInitializer : Initializer<Unit> {
     override fun create(context: Context) {
